+++ Source/DataStructureAndEncodingDefinition/gdcmReader.cxx
@@ -433,10 +433,10 @@ bool Reader::InternalReadCommon(const T_Caller &caller)
           {
           std::streampos start = is.tellg();
           is.seekg( 0, std::ios::end);
-          std::streampos end = is.tellg();
-		  std::streamoff theOffset = end-start;
-		  assert (theOffset > 0 || theOffset < std::numeric_limits<uint32_t>::max());
-          VL l = (uint32_t)(theOffset);
+	  std::streampos end = is.tellg();
+	  std::streamoff theOffset = end-start;
+	  assert (theOffset > 0 || (uint32_t)theOffset < std::numeric_limits<uint32_t>::max());
+	  VL l = (uint32_t)(theOffset);
           is.seekg( start, std::ios::beg );
           //F->GetDataSet().ReadUpToTagWithLength<ImplicitDataElement,SwapperNoOp>(is, tag, l);
           caller.template ReadCommonWithLength<ImplicitDataElement,SwapperNoOp>(is,l);
